+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Removed chat popup.
+- Made excerpt not mandatory for posts.
 
 ## 0.1.0